% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{A simple dataset}
\format{A data frame with 10 rows and 2 variables:
\describe{
  \item{X}{integers 1 to 10}
  \item{Y}{the square of X}
  ...
}}
\usage{
example_data
}
\description{
A simple dataset
}
\keyword{datasets}
