% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_caption.R
\name{figure.caption}
\alias{figure.caption}
\title{Figure Caption}
\usage{
figure.caption(caption, caption.width = 98)
}
\arguments{
\item{caption}{A string containing the caption text.}

\item{caption.width}{A positive integer giving the target column for wrapping lines in the output. Defaults to 98}
}
\description{
Create an auto-numbered, wrapped figure caption
}
\examples{
data(example_data)

plot(example.data$X, example.data$Y, sub = figure.caption("This is a figure"))
}
\seealso{
\code{\link{start.tracking}}, \code{\link{wrap.caption}}, \code{\link{supp.fig.caption}}
}
