% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracking_num.R
\name{get.tracking.num}
\alias{get.tracking.num}
\title{Get Figure/Table Number}
\usage{
get.tracking.num(type = "figure")
}
\arguments{
\item{type}{A character vector indicating whether you want the current "figure", "table", "supplemental figure", or "supplemental table" number. Defaults is "figure".}
}
\description{
Get current figure, table, supplemental figure, and/or supplemental table number
}
\examples{
data(example_data)

plot(example.data$X, example.data$Y, sub = figure.caption("This is a figure"))

get.tracking.num(type = "figure")

plot(example.data$X, example.data$Y, sub = figure.caption("This is another figure"))
}
