% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_figure_caption.R
\name{gg_figure_caption}
\alias{gg_figure_caption}
\title{ggplot Figure Caption}
\usage{
gg_figure_caption(
  caption,
  supplemental = FALSE,
  s.numbers = FALSE,
  caption.width = 98
)
}
\arguments{
\item{caption}{A string containing the caption text.}

\item{supplemental}{Is this a supplemental figure? Defaults to FALSE.}

\item{s.numbers}{If it is a supplmental figure, how to you want it formatted: "Supplemental Figure 1" (FALSE) or "Figure S1" (TRUE)? Defaults to FALSE.}

\item{caption.width}{A positive integer giving the target column for wrapping lines in the output. Defaults to 98}
}
\description{
Add an auto-numbered caption to a ggplot.
}
\examples{
data(example_data)

ggplot(example.data, aes(x = X, y = Y)) +
  geom_point() +
  gg_figure_caption(caption = "This is a figure")
}
\seealso{
\code{\link{ggplot2::labs}}, \code{\link{figure.caption}}, \code{\link{supp.fig.caption}}
}
