% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tracking_nums.R
\name{set.tracking.nums}
\alias{set.tracking.nums}
\title{Set Figure/Table Numbers}
\usage{
set.tracking.nums(
  type = c("figure", "table", "supplemental figure", "supplemental table"),
  set.to = 0
)
}
\arguments{
\item{type}{A character vector indicating whether you want figure, table, supplemental figure, and/or supplemental table numbers set to specific values. Defaults to all four: c("figure", "table", "supplmental figure", "supplemental table")}

\item{set.to}{A positive integer vector giving the the new values you want set for figure/table numbers. Must be in same order as the type vector, or a single value to set them all to. Defaults to 0.}
}
\description{
Set figure, table, supplemental figure, and/or supplemental table numbers to a specific value
}
\examples{
data(example_data)

plot(example.data$X, example.data$Y, sub = figure.caption("This is a figure"))

set.tracking.nums(type = "figure", set.to = 3)

plot(example.data$X, example.data$Y, sub = figure.caption("This is another figure"))
}
