% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_fig_caption.R
\name{supp.fig.caption}
\alias{supp.fig.caption}
\title{Supplemental Figure Caption}
\usage{
supp.fig.caption(caption, s.numbers = FALSE, caption.width = 98)
}
\arguments{
\item{caption}{A string containing the caption text.}

\item{s.numbers}{If it is a supplmental figure, how to you want it formatted: "Supplemental Figure 1" (FALSE) or "Figure S1" (TRUE)? Defaults to FALSE.}

\item{caption.width}{A positive integer giving the target column for wrapping lines in the output. Defaults to 98}
}
\description{
Create an auto-numbered, wrapped supplemental figure caption
}
\examples{
data(example_data)

plot(example.data$X, example.data$Y, sub = supp.fig.caption("This is a figure"))
}
\seealso{
\code{\link{figure.caption}}, \code{\link{wrap.caption}}
}
