% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_tbl_caption.R
\name{supp.tbl.caption}
\alias{supp.tbl.caption}
\title{Supplemental Table Caption}
\usage{
supp.tbl.caption(caption, s.numbers = FALSE, caption.width = 98)
}
\arguments{
\item{caption}{A string containing the caption text.}

\item{s.numbers}{If it is a supplmental figure, how to you want it formatted: "Supplemental Table 1" (FALSE) or "Table S1" (TRUE)? Defaults to FALSE.}

\item{caption.width}{A positive integer giving the target column for wrapping lines in the output. Defaults to 98}
}
\description{
Create an auto-numbered, wrapped supplemental table caption
}
\examples{
data(example_data)

knitr::kable(example.data, format = "latex", caption = supp.tbl.caption("This is a table"))
}
\seealso{
\code{\link{start.tracking}}, \code{\link{table.caption}}, \code{\link{wrap.caption}}
}
