% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_env.R
\docType{data}
\name{tracking.env}
\alias{tracking.env}
\title{Tracking Environment}
\format{An object of class \code{environment} of length 4.}
\usage{
tracking.env
}
\description{
Create environment for tracking figures and tables.
}
\keyword{datasets}
