% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_caption.R
\name{wrap.caption}
\alias{wrap.caption}
\title{Wrap Caption}
\usage{
wrap.caption(caption, w = 98)
}
\arguments{
\item{caption}{A string containing the caption text.}

\item{w}{A positive integer giving the target column for wrapping lines in the output. Defaults to 98}
}
\description{
Adds line breaks at the appropriate spots in a caption string
}
\seealso{
\code{\link{strwrap}}, \code{\link{paste}}
}
